+++ src/util/small_map.h
@@ -22,6 +22,7 @@ Author: Daniel Poetzl
 #include <utility>
 
 #include "invariant.h"
+#include "narrow.h"
 
 //#define _SMALL_MAP_REALLOC_STATS
 
@@ -380,7 +381,7 @@ public:
 
   const_value_iterator value_begin() const
   {
-    return const_value_iterator(*this, size() - 1);
+    return const_value_iterator(*this, narrow_cast<int>(size()) - 1);
   }
 
   const_value_iterator value_end() const
